pkg/endpoint: move status-related code into status.go
This file already contained status-related information, put functions with
endpoint receiver into it as well.

Signed-off by: Ian Vernon <ian@cilium.io>
